requirements.txt
# Unified Monitoring Automation Requirements
# Both Grafana and Humio automation dependencies

# Core dependencies
playwright==1.49.0          # Browser automation (async/sync APIs)
reportlab==4.2.5            # PDF report generation
Pillow>=10.0.0              # Image processing for screenshots
python-dotenv>=1.0.0        # Environment variable management
